refactor(ui): move Result Config to Results page
- Add result_config field to TournamentDetail schema and backend response
- Remove Result Config section from TournamentEditPage
- Add Result Config form (organizer-only) to ResultsPage, pre-populated from tournament data
- Remove onSuccess invalidation from useUpdateResultConfig to prevent form state reset
- Remove /uploads/{key} from OpenAPI paths (served directly by Gin, must not be code-generated)

Closes #132
